koinモジュール定義の更新
